backend/requirements.txt
scikit-learn==1.6.1
flask==3.0.0
flask-cors==4.0.0
flask-jwt-extended==4.6.0
flask-sqlalchemy==3.1.1
bcrypt==4.1.2
pandas==2.2.0
numpy==1.26.4
joblib==1.3.2
spacy==3.7.2
nltk==3.8.1